// src/app/api/models/matricula-estado-cuenta.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PagoResponse } from '../models/pago-response';
export interface MatriculaEstadoCuenta {
  estado?: 'PENDIENTE' | 'PAGADO' | 'ANULADO' | 'VENCIDO';
  id?: number;
  montoTotal?: number;
  morosidad?: number;
  pagos?: Array<PagoResponse>;
  periodoAcademico?: string;
  saldoPendiente?: number;
  totalPagado?: number;
}