// src/app/api/models/pago-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PagoResponse {
  fechaPago?: string;
  id?: number;
  idAlumno?: number;
  idCuota?: number;
  idMatricula?: number;
  metodoPago?: 'EFECTIVO' | 'TARJETA_DEBITO' | 'TARJETA_CREDITO' | 'TRANSFERENCIA_BANCARIA' | 'YAPE' | 'PLIN';
  montoPagado?: number;
  numeroOperacion?: string;
  observaciones?: string;
  pasarela?: 'Culqui';
}